# rentalApp/src/main/resources/app-api-swagger.json.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/pet/booking/{petId}": {
      "get": {
        "tags": [
          "Booking Controller"
        ],
        "summary": "Booking form",
        "operationId": "showBookingForm",
        "parameters": [
          {
            "name": "userDetails",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserDetails"
            }
          },
          {
            "name": "petId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "bookingCreateEditDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BookingCreateEditDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Booking Controller"
        ],
        "summary": "Order pet",
        "operationId": "makeOrder",
        "parameters": [
          {
            "name": "petId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "duration",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "booking": {
                    "$ref": "#/components/schemas/BookingCreateEditDto"
                  },
                  "userBooking": {
                    "$ref": "#/components/schemas/UserCreateEditDto"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/update-user/{id}": {
      "get": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "User data",
        "operationId": "showUserDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Update user data",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateEditDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/update-pet/{id}": {
      "get": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Pet details",
        "operationId": "showPetDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Update pet data",
        "operationId": "updatePet",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "petCreateEditDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PetCreateEditDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/listOfBookingDetails/{id}": {
      "get": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Booking details specific user",
        "operationId": "showBookingDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Update user status of booking",
        "operationId": "updateBookingStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreateEditDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/pet/bookings": {
      "get": {
        "tags": [
          "Booking Controller"
        ],
        "summary": "Bookings by user",
        "operationId": "getBookings",
        "parameters": [
          {
            "name": "userDetails",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserDetails"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/pets-info": {
      "get": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Find all pets",
        "operationId": "findAllPets",
        "parameters": [
          {
            "name": "petFilter",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PetFilter"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/medical-record-info/{petId}": {
      "get": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Pet medical records",
        "operationId": "medicalInfo",
        "parameters": [
          {
            "name": "petId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/listOfBookings": {
      "get": {
        "tags": [
          "Admin Controller"
        ],
        "summary": "Find all bookings by app users",
        "operationId": "findAllBookings",
        "parameters": [
          {
            "name": "bookingStatus",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "AwaitingApproval",
                "ReadyToDiscuss",
                "InProcess",
                "RentalEnded",
                "CANCELED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserCreateEditDto": {
        "required": [
          "address",
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "maxLength": 20,
            "minLength": 8,
            "pattern": "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z]).{8,}$",
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "firstname": {
            "maxLength": 20,
            "minLength": 2,
            "pattern": "[A-Za-z]+",
            "type": "string"
          },
          "lastname": {
            "maxLength": 25,
            "minLength": 2,
            "pattern": "[A-Za-z]+",
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "pesel": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "profilePicture": {
            "type": "string",
            "format": "binary"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "emailVerified": {
            "type": "boolean"
          },
          "emailVerificationToken": {
            "type": "string"
          }
        }
      },
      "BookingCreateEditDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "pet": {
            "$ref": "#/components/schemas/Pet"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "AwaitingApproval",
              "ReadyToDiscuss",
              "InProcess",
              "RentalEnded",
              "CANCELED"
            ]
          }
        }
      },
      "Clinic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "veterinarians": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Veterinarian"
            }
          }
        }
      },
      "MedicalRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "pet": {
            "$ref": "#/components/schemas/Pet"
          },
          "veterinarian": {
            "$ref": "#/components/schemas/Veterinarian"
          },
          "diagnosis": {
            "type": "string"
          },
          "treatment": {
            "type": "string"
          },
          "prescription": {
            "type": "string"
          },
          "examinationDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "Pet": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "species": {
            "$ref": "#/components/schemas/Species"
          },
          "breed": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE"
            ]
          },
          "medicalRecords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicalRecord"
            }
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "isAvailable": {
            "type": "boolean"
          }
        }
      },
      "Species": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "pesel": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "profilePicture": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "emailVerified": {
            "type": "boolean"
          },
          "emailVerificationToken": {
            "type": "string"
          }
        }
      },
      "Veterinarian": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "clinic": {
            "$ref": "#/components/schemas/Clinic"
          },
          "vetCode": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "specialization": {
            "type": "string"
          },
          "contactNumber": {
            "type": "string"
          }
        }
      },
      "MedicalRecordReadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "pet": {
            "$ref": "#/components/schemas/Pet"
          },
          "veterinarian": {
            "$ref": "#/components/schemas/Veterinarian"
          },
          "diagnosis": {
            "type": "string"
          },
          "treatment": {
            "type": "string"
          },
          "prescription": {
            "type": "string"
          },
          "examinationDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "PetCreateEditDto": {
        "type": "object",
        "properties": {
          "species": {
            "$ref": "#/components/schemas/Species"
          },
          "breed": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE"
            ]
          },
          "medicalRecords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedicalRecordReadDto"
            }
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "isAvailable": {
            "type": "boolean"
          }
        }
      },
      "GrantedAuthority": {
        "type": "object",
        "properties": {
          "authority": {
            "type": "string"
          }
        }
      },
      "UserDetails": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "username": {
            "type": "string"
          },
          "authorities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrantedAuthority"
            }
          },
          "accountNonExpired": {
            "type": "boolean"
          },
          "credentialsNonExpired": {
            "type": "boolean"
          },
          "accountNonLocked": {
            "type": "boolean"
          }
        }
      },
      "PetFilter": {
        "type": "object",
        "properties": {
          "species": {
            "$ref": "#/components/schemas/Species"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE"
            ]
          },
          "isAvailable": {
            "type": "boolean"
          }
        }
      },
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}